Fixed entities moving to wrong coordinates after their world is removed
When a player world was removed, it was removed as a component from that
entity too. Then, when it was reassigned a world there was no way to
determine the delta translation. Now we keep that world until after it's
reassigned to a new world, so that we can calculate the delta and move
it appropriately.